Month2/Week8/Day5_15_Jan_2026/rate_limit_api/requirements.txt
fastapi
uvicorn
redis
python-jose
pydantic